+++ src/com/oDesk/api/Routers/Activities/Engagement.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.oDesk.api.Routers.Hr.Clients;
+package com.oDesk.api.Routers.Activities;
 
 import java.util.HashMap;
 
@@ -32,36 +32,40 @@ import org.json.JSONObject;
 	lastModifiedBy = "Maksym Novozhylov",
 	reviewers = {"Yiota Tsakiri"}
 )
-public final class Applications {
+public final class Engagement {
 	
 	final static String ENTRY_POINT = "api";
 	
 	private OAuthClient oClient = null;
 
-	public Applications(OAuthClient client) {
+	public Engagement(OAuthClient client) {
 		oClient = client;
 		oClient.setEntryPoint(ENTRY_POINT);
 	}
 	
 	/**
-     * Get list of applications
+     * List activities for specific engagement
      *
-     * @param   params Parameters
-     * @return  object
+     * @param   engagement_ref Engagement reference
+     * @throws	JSONException If error occurred
+	 * @return	{@link JSONObject}
      */
-    public JSONObject getList(HashMap<String, String> params) throws JSONException {
-        return oClient.get("/hr/v3/clients/applications", params);
+    public JSONObject getSpecific(String engagement_ref) throws JSONException {
+        return oClient.get("/tasks/v2/tasks/contracts/" + engagement_ref);
     }
-
+    
     /**
-     * Get specific application
+     * Assign engagements to the list of activities
      *
-     * @param   reference Application reference
+     * @param   company Company ID
+     * @param	team Team ID
+     * @param   engagement Engagement
      * @param   params Parameters
-     * @return  object
+     * @throws	JSONException If error occurred
+	 * @return	{@link JSONObject}
      */
-    public JSONObject getSpecific(String reference, HashMap<String, String> params) throws JSONException {
-        return oClient.get("/hr/v3/clients/applications/" + reference, params);
+    public JSONObject assign(String company, String team, String engagement, HashMap<String, String> params) throws JSONException {
+        return oClient.put("/otask/v1/tasks/companies/" + company + "/teams/" + team + "/engagements/" + engagement + "/tasks", params);
     }
-
+    
 }